add from_scalar method
