+++ mcl-kzg/kzg/src/kzg_settings.rs
@@ -1,4 +1,4 @@
-use crate::data_types::{g1::G1, g2::G2};
+use crate::data_types::{fr::Fr, g1::G1, g2::G2};
 use crate::fk20_fft::FFTSettings;
 
 #[derive(Clone)]
@@ -22,6 +22,23 @@ impl KZGSettings {
             secret1: sec1,
             secret2: sec2
         }
+    }
+    pub fn generate_trusted_setup(n: usize, secret: [u8; 32usize]) -> (Vec<G1>, Vec<G2>) {
+        let g1_gen = G1::gen();
+        let g2_gen = G2::gen(); 
+
+        let mut g1_points = vec!(G1::default(); n);
+        let mut g2_points = vec!(G2::default(); n);
+        let secretfr = Fr::from_scalar(&secret);
+        let mut secret_to_power = Fr::one();
+        for i in 0..n {
+            G1::mul(&mut (g1_points[i]), &g1_gen, &secret_to_power);
+            G2::mul(&mut (g2_points[i]), &g2_gen, &secret_to_power);
+
+            secret_to_power *= &secretfr;
+        }
+
+        (g1_points, g2_points)
     }
 }
